feat: Provider new structure, show transaction fee
